Adopt phonecall behaviour from earlier version.
